backend/requirements.txt
Flask==2.0.1
Flask-Pymongo==2.3.0
transformers==4.28.0
torch==2.7.1
python-dotenv==1.0.1
flask-cors